requirements.txt
boto3
h5py
pyyaml